% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_groups}
\alias{animal_groups}
\alias{animal_traits}
\title{Small data set of animals.}
\format{
\code{animal_groups}: A MultiFactor of animals as well as associated
\code{emoji}.

\code{animal_traits}: A MultiFactor of animals as well as associated
\code{emoji}.
}
\source{
\code{animal_groups}: Manually curated.

\code{animal_traits}: Manually curated.
}
\usage{
data("animals", package = "MultiFactor")

data("animals", package = "MultiFactor")
}
\description{
Small data set of animals.

Small data set of animals.
}
\keyword{datasets}
