% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_conversion_utils.R
\name{as.igraph.LinkMap}
\alias{as.igraph.LinkMap}
\alias{as.igraph.MultiFactor::LinkMap}
\title{Convert a LinkMap to igraph}
\usage{
\method{as.igraph}{`MultiFactor::LinkMap`}(x, directed = FALSE, ...)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{directed}{See \code{?igraph::graph_from_biadjacency_matrix}}

\item{...}{Additional arguments passed to
\code{igraph::graph_from_biadjacency_matrix}.}
}
\value{
an \code{igraph} object.
}
\description{
Extract relational information from LinkMap and return an igraph object.
Calls \code{igraph::graph_from_biadjacency_matrix} under the hood.
}
\details{
If a \code{LinkMap} contains metadata, this can be found back as edge
attributes in the returned graph (see \code{igraph::edge_attr()})
}
\examples{
x <- randomLinkMap()

# Make igraph object:
igraph::as.igraph(x)

}
