% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.LinkMap.R
\name{as.LinkMap-methods}
\alias{as.LinkMap-methods}
\title{Convert common classes to LinkMap or MultiFactor}
\arguments{
\item{x}{\code{data.frame}, \code{list} or \code{logical matrix}. Input
object(s) to convert to \code{LinkMap} or \code{MultiFactor}.}

\item{y}{\code{Character vector}. The vector of elements to use as
feature names along with \code{x} if the latter is an unnamed list.
(Default: \code{NULL}).}

\item{edge.names}{\code{Character vector}. A vector of two elements
specifying the names of the edges in the output \code{LinkMap}.
(Default: \code{NULL}).}
}
\value{
\code{as.LinkMap} returns a \code{LinkMap} where each row contains a unique
combination of the elements in \code{x} (and \code{y} for an unnamed list).
\code{as.MultiFactor} returns a \code{MultiFactor}
}
\description{
\code{as.LinkMap} and \code{as.MultiFactor }convert the input object(s) to a
\code{LinkMap} and a \code{MultiFactor}, respectively.
}
\examples{
# Create list with random linkage
lst <- lapply(1:3, function(x) sample(letters[seq(3)], 3, replace = TRUE))
names(lst) <- LETTERS[seq(3)]

# Convert to LinkMap
a2b <- as.LinkMap(lst, edge.names = c("a", "b"))

# Create adjacency matrix
mat <- matrix(sample(c(TRUE, FALSE), 3 * 3, replace = TRUE), nrow = 3)
rownames(mat) <- LETTERS[seq(3)]
colnames(mat) <- c("x", "y", "z")

# Convert adjacency matrix to LinkMap
b2c <- as.LinkMap(mat, edge.names = c("b", "c"))

# Convert list of lists, matrices or data.frames to MultiFactor
mf <- as.MultiFactor(list(X1 = lst, X2 = mat))

}
