% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkMap-convert.R
\name{as.linkmap}
\alias{as.linkmap}
\title{Convert to LinkMap}
\usage{
as.linkmap(values, keys = NULL, col.names = NULL)
}
\arguments{
\item{values}{\code{Character list}. List of vectors where each vector
represents one type of values.}

\item{keys}{\code{Character vector}. The vector of keys to use as names for
\code{values} if the latter is unnamed. (Default: \code{NULL}).}

\item{col.names}{\code{Character vector}. A vector of two elements specifying
the names of the columns in the output linkmap (Default: \code{NULL}).}
}
\value{
\code{as.linkmap} returns a linkmap \code{data.frame} where the first and
second columns contains \code{keys} and \code{values} and each row represents
a unique combination of the two.
}
\description{
\code{as.linkmap} converts a list of named vectors to a linkmap data.frame.
}
\examples{
# Create list with random linkage
a2b <- lapply(1:3, function(x) sample(letters[seq(3)], 3, replace = TRUE))
names(a2b) <- LETTERS[seq(3)]

# Convert to LinkMap
as.linkmap(a2b)
}
