% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-subset.R
\name{augment-subset}
\alias{augment-subset}
\alias{augment.MultiFactor}
\alias{subset.MultiFactor::MultiFactor}
\alias{subset.LinkMap}
\alias{subset.MultiFactor::LinkMap}
\alias{augment.MultiFactor::MultiFactor}
\title{Tools to modify MultiFactors}
\usage{
\method{subset}{`MultiFactor::MultiFactor`}(x, .path, .drop.unmatched = FALSE, ...)

\method{subset}{`MultiFactor::LinkMap`}(x, subset = NULL, ...)

\method{augment}{`MultiFactor::MultiFactor`}(x, ...)
}
\arguments{
\item{x}{a \code{MultiFactor} or \code{LinkMap}.}

\item{.path}{Either a \code{formula} or a \verb{character vector}.}

\item{.drop.unmatched}{\code{Boolean}. Whether to drop factor levels that do not
appear in the data (Default: keep all levels).}

\item{...}{Name-value pairs. The name gives the name of the LinkMap in the
output.}

\item{subset}{Logical expression indicating elements or rows to keep.
If \code{NULL} (Default), subsets based on complete coverage if available.}
}
\value{
a \code{MultiFactor}.
}
\description{
Generates a new \code{MultiFactor} object by cross-referencing the elements of a
given \code{MultiFactor}.
}
\examples{
# Only necessary in example code
require(generics)
# Generate a random MultiFactor
x <- randomMultiFactor()

# Use augment to chain together operations like weave and stack, in order.
augment(x,
    weave(x, a ~ c),
    stack(x, a + b ~ c + d),
    weave(x, d ~ f)
)

# Setting a LinkMap to NULL by name deletes it from the MultiFactor
augment(x, a2b = NULL )

}
