% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-utils.R
\name{.set_levels_MultiFactor}
\alias{.set_levels_MultiFactor}
\title{Set levels of a MultiFactor Object}
\usage{
.set_levels_MultiFactor(x, levels)
}
\arguments{
\item{x}{MultiFactor or appropriately formatted list}

\item{levels}{A \verb{named list of character vectors}, to be used as
replacement levels.}
}
\value{
a MultiFactor with updated levels.
}
\description{
Set levels of a MultiFactor Object
}
