% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{factor_path}
\alias{factor_path}
\title{factor_path S7 class}
\usage{
factor_path(
  x,
  include = list(character()),
  exclude = character(),
  exact = FALSE
)
}
\arguments{
\item{x}{\verb{Character vector} of length two; \verb{c(<from>, <to>)}.}

\item{include}{\verb{List of character vectors}. Which elements are included in
the path? Length of the list indicates number of paths.}

\item{exclude}{\verb{Character vector}. Which elements are excluded from the path?}

\item{exact}{\code{Logical}. Should the path be followed as is?
(i.e., no pathfinding required)}
}
\value{
a \code{factor_path} object.
}
\description{
\code{factor_path} is an S7 class to standardize the various ways in which paths
that can be specified across the \code{MultiFactor} package. This class is mostly
for internal use.
}
\examples{
factor_path(x = c("a", "c"), include = list("b"))
}
