% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_path-methods.R
\name{factor_path-methods}
\alias{factor_path-methods}
\title{Methods for factor_path S7 class}
\arguments{
\item{x, object}{\code{factor_path} on which the method should be applied.}
}
\value{
A \code{factor_path}
}
\description{
Methods for factor_path S7 class
}
\examples{
# Setup

set.seed(2612)
tp <- trade_posts()
x <- select_path(tp, fruit ~ furniture)

# Basic properties
length(x)
terms(x)

}
