% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexLinkMap.R
\name{groupwise_apply}
\alias{groupwise_apply}
\title{Index a table and apply arbitrary code to it}
\usage{
groupwise_apply(X, LINK, BY, FUN, ...)
}
\arguments{
\item{X}{\verb{A table}. A \code{data.frame}, \code{matrix}, other object with rows and
columns.}

\item{LINK}{a \code{MultiFactor} object.}

\item{BY}{either a \code{formula} or a `character vector`` of length 2 with the
names of the desired combination of feature types.}

\item{FUN}{A function, passed to lapply.}

\item{...}{Additional arguments passed to lapply call.}
}
\value{
a \verb{Named list} of desired output.
}
\description{
\code{groupwise_apply} can run arbitrary code specified by \code{FUN} across any number
of row-based subsets of table \code{X}. Mimics lapply.
}
\examples{

# Prepare data
link <- anansi::kegg_link()
data("FMT_data", package = "anansi")
x <- FMT_KOs

# Apply arbitrary code to x based on group membership
groupwise_apply(x, link, BY =  ec ~ ko, FUN = function(x) dim(x))

# Utilitties

.index_tbl_by(x, link, ec ~ ko)

# Any table can be a indexed into a LinkMap:
.index_tbl(mtcars)

}
\seealso{
\code{\link[=weave]{weave()}} \code{\link[=LinkMap]{LinkMap()}} \code{\link[=MultiFactor]{MultiFactor()}}
}
