% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexLinkMap.R
\name{.index_tbl_by}
\alias{.index_tbl_by}
\alias{.index_tbl}
\title{Index a table}
\usage{
.index_tbl(X, type = deparse1(substitute(X)), ...)

.index_tbl_by(X, link, .by, ...)
}
\arguments{
\item{type}{\verb{Character scalar} Specifies name of the feature type, will be
used as name of the first column in output.}

\item{link}{a \code{MultiFactor} object.}

\item{.by}{either a \code{formula} or a `character vector`` of length 2 with the
names of the desired combination of feature types.}
}
\description{
\code{.index_tbl_by()} finds row indices for input table \code{x}, based on a
user-defined path through a MultiFactor (\code{link} arg).

\code{.index_tbl()} is a utility function that flexibly takes a table as input and
returns a special \code{LinkMap} object that has (1) row names and (2)
corresponding row indices of that table as columns.
}
