% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexLinkMap.R
\name{indexLinkMap}
\alias{indexLinkMap}
\title{Index a table}
\usage{
indexLinkMap(x, type = deparse1(substitute(x)), ...)
}
\arguments{
\item{x}{\verb{A table}. A \code{data.frame}, \code{matrix}, other object with rows and
columns.}

\item{type}{\verb{Character scalar} Specifies name of the feature type, will be
used as name of the first column in output.}

\item{...}{Additional arguments passed to \code{LinkMap()} call.}
}
\value{
a \code{LinkMap} object.
}
\description{
\code{indexLinkMap} is a utility function that flexibly takes a table as input and
returns a special \code{LinkMap} object that has (1) row names and (2)
corresponding row indices of that table as columns. It can be used with
\code{weave()} to obtain row indices in based on possibly overlapping groupings.
}
\examples{
# Any table can be a indexed
indexLinkMap(mtcars)

}
\seealso{
\code{\link[=LinkMap]{LinkMap()}} \code{\link[=MultiFactor]{MultiFactor()}}
}
