% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkMap-methods.R
\name{LinkMap-methods}
\alias{LinkMap-methods}
\title{Methods for LinkMap S7 container class}
\arguments{
\item{x, object}{\code{LinkMap} on which the method should be applied.}
}
\value{
A \code{LinkMap}
}
\description{
Methods for LinkMap S7 container class
}
\examples{
# Setup
a2b <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)

# Create LinkMap
x <- LinkMap(a2b)

# Basic properties
dim(x)
dimnames(x)

# Factor-like properties
levels(x)
nlevels(x)


}
