% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_conversion_utils.R
\name{mf_as_graph_df}
\alias{mf_as_graph_df}
\title{Convert a MultiFactor to relational graph format.}
\usage{
mf_as_graph_df(x)
}
\arguments{
\item{x}{MultiFactor}
}
\value{
a \code{data.frame} with three or more named columns; from, to and name.
}
\description{
Utility function to convert graph information of a MultiFactor to a
\code{data.frame}. Used in \code{as.igraph} method for MultiFactor.
}
\examples{
x <- randomMultiFactor()

# Three-column dfs of layouts
x_df <- mf_as_graph_df(x)

}
