% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_conversion_utils.R
\name{mf_to_graph_df}
\alias{mf_to_graph_df}
\title{Convert a MultiFactor to relational graph format.}
\usage{
mf_to_graph_df(x, source_name)
}
\arguments{
\item{x}{MultiFactor}

\item{source_name}{\verb{Character scalar} Content of source column in output.}
}
\value{
a \code{data.frame} with three named columns; from, to and source_name.
}
\description{
Utility function to convert graph information of a MultiFactor to a
\code{data.frame}. Used in \code{as.igraph} method for MultiFactor.
}
\examples{
x <- randomMultiFactor()

# Three-column dfs of layouts
x_df <- mf_to_graph_df(x, source_name = "example_name")

}
