% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor-wrangle-utils.R
\name{MultiFactor-wrangle-methods}
\alias{MultiFactor-wrangle-methods}
\alias{augment.MultiFactor}
\alias{augment.MultiFactor::MultiFactor}
\title{Tools to modify MultiFactors}
\usage{
\method{augment}{`MultiFactor::MultiFactor`}(x, ...)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{...}{Name-value pairs. The name gives the name of the LinkMap in the
output.}
}
\value{
a \code{MultiFactor}.
}
\description{
Generates a new \code{MultiFactor} object by cross-referencing the elements of a
given \code{MultiFactor}.
}
\examples{
# Only necessary in example code
require(generics)
# Generate a random MultiFactor
x <- randomMultiFactor()

# Use augment to build upon the same MultiFactor.
augment(x,
    weave(x, a ~ c),
    weave(x, d ~ f)
)
# Setting a LinkMap to NULL deletes it from the MultiFactor
augment(x, a2b = NULL )

}
