% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nlevels}
\alias{nlevels}
\title{The Number of Levels of an Object}
\usage{
nlevels(x, ...)
}
\arguments{
\item{x}{an object, such as a \code{LinkMap}, \code{MultiFactor} or \code{factor}.}

\item{...}{additional arguments. Not used for \code{base::factor} method.}
}
\value{
A \verb{Numeric vector} of length equal to the number of elements in \code{x}.
Optionally, named.
}
\description{
Return the number of levels which its argument has. Extends \code{base::nlevels}.
}
\examples{
# Available methods:
nlevels

}
