% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{playing_cards}
\alias{playing_cards}
\title{Basic names of entities in a card deck.}
\format{
\code{playing_cards}: A list with two character vectors; contains the
four \code{suits} and thirteen \code{ranks} of a standard poker deck. Used in the
vignette.
}
\source{
\code{playing_cards}: Manual input.
}
\usage{
data("playing_cards", package = "MultiFactor")
}
\description{
Basic names of entities in a card deck.
}
\keyword{datasets}
