% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/quick_brown_fox_data.R
\docType{data}
\name{quickbrownfox}
\alias{quickbrownfox}
\title{Example dataset using a famous pangram.}
\format{
\code{quickbrownfox}: A MultiFactor of letters, words and constituents
}
\source{
\code{quickbrownfox}: Manually curated.
}
\usage{
data("quickbrownfox", package = "MultiFactor")

quickbrownfox(raw.data = FALSE)
}
\arguments{
\item{raw.data}{\code{Boolean}, Whether to return underlying \code{list} rather
than the default \code{MultiFactor}.}
}
\description{
Demo data set used in the coverage en enrichment vignette.
}
\examples{
quickbrownfox()

}
\keyword{datasets}
