% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-utils.R
\name{read_adjacency_list}
\alias{read_adjacency_list}
\title{Read Relational Data Values}
\usage{
read_adjacency_list(
  file,
  sep = "\\t",
  quote = "\\"'",
  col.names = c("id.x", "id.y"),
  as.df = FALSE,
  ...
)
}
\arguments{
\item{file}{the name of a file to read data values from.  If the
    specified file is \code{""}, then input is taken from the keyboard
    (or whatever \code{\link[base]{stdin}()} reads if input is redirected or
    \R is embedded).
    (In this case input can be terminated by a blank line or an \abbr{EOF}
    signal, \samp{Ctrl-D} on Unix and \samp{Ctrl-Z} on Windows.)

    Otherwise, the file name is interpreted \emph{relative} to the
    current working directory (given by \code{\link[base]{getwd}()}),
    unless it specifies an \emph{absolute} path.
    Tilde-expansion is performed where supported.
    When running \R from a script, \code{file = "stdin"} can be used to
    refer to the process's \code{stdin} file stream.

    This can be a compressed file (see \code{\link[base]{file}}).

    Alternatively, \code{file} can be a \code{\link[base]{connection}},
    which will be opened if necessary, and if so closed at the end of
    the function call.  Whatever mode the connection is opened in,
    any of \abbr{LF}, \abbr{CRLF} or \abbr{CR} will be accepted as the
    \abbr{EOL} marker for a line and so will match \code{sep = "\n"}.

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link[base]{url}}.)

    To read a data file not in the current encoding (for example a
    Latin-1 file in a UTF-8 locale or conversely) use a
    \code{\link[base]{file}} connection setting its \code{encoding} argument
    (or \code{scan}'s \code{fileEncoding} argument).
  }

\item{sep}{the field separator character. Values on each line of the file
are separated by this character. (Default: \code{"\\t"})}

\item{quote}{the set of quoting characters as a single character
    string or \code{NULL}.  In a multibyte locale the quoting characters
    must be ASCII (single-byte).}

\item{col.names}{\verb{Character vector} of length 2, specifying the colnames of
the output. (Default: \code{c("id.x", "id.y")})}

\item{as.df}{\code{Boolean} Whether to return an unmodified \code{data.frame} or a
\code{LinkMap} (Default).}

\item{...}{Additional arguments, passed to \code{scan}.}
}
\value{
A \code{LinkMap}, or \verb{if( as.df )}, a two-column data.frame formatted
like an edge list, appropriate input for \code{LinkMap()}.
}
\description{
Read relational data from a file in adjacency list format and parse it into a
\code{LinkMap} (or a \code{LinkMap}-shaped \code{data.frame}). Relies on \code{base::scan} and
\code{utils::count.fields} under the hood.
}
\details{
Note that the other common representation of relational data, a two-column
table, known as an edge list, can be read in using regular approaches
(e.g., \code{?base::read.table}).
}
\examples{
# read_adjacency_list(
#     "https://ftp.microbio.me/pub/wol2/function/kegg/ko-to-ec.map"
# )

}
\seealso{
\code{\link[utils:count.fields]{utils::count.fields()}} \code{\link[base:scan]{base::scan()}}
}
