% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{select_path-generic}
\alias{select_path-generic}
\alias{select_path}
\title{Weave a path through an object}
\usage{
select_path(x, .path, ...)
}
\arguments{
\item{x}{input object}

\item{.path}{either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{...}{additional arguments}
}
\value{
a (list of) character vector(s).
}
\description{
\code{select_path()} is an S7 generic that finds and returns a path
through a relational object.
}
\examples{
# Available methods:
select_path

}
