% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_path.R
\name{select_path-methods}
\alias{select_path-methods}
\title{Define a path through a MultiFactor object.}
\arguments{
\item{x}{input object}

\item{.path}{either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}
}
\value{
a \code{factor_path} object.
}
\description{
Define a path through a MultiFactor object.
}
\examples{
#' # Generate pair of random linkage input
a2b <- data.frame(
   a = sample(letters[seq(3)], 10, replace = TRUE),
   b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
a2c <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    c = sample(c("x", "y", "z"), 10, replace = TRUE)
)

# Create MultiFactor
x <- MultiFactor(list(a2b, a2c))

# Inspect a path between data types
select_path(x, b ~ c)

}
