% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack-methods}
\alias{stack-methods}
\title{Combine levels across several LinkMaps in a MultiFactor}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.path}{a \code{formula} of length 2 with with levels to be merged separated
by the plus (\code{+}) sign. Optionally, a list with two character vectors,
signifying the variables to be combined at the left and right hand side,
respectively.}

\item{...}{Additional arguments (unused.)}

\item{out.format}{\verb{Character scalar}. One of \code{'LinkMap'}, \code{'matrix'}.}
}
\value{
a \code{LinkMap} or \verb{sparse Matrix}.
}
\description{
Generates a new \code{LinkMap} object by merging levels by name,
separated by the plus (\code{+}) sign. See examples.
}
\examples{
# Only necessary in example code
require(utils)

x <- randomMultiFactor()
# Merge variables with "+" operator, new names get concatenated with ".":
stack(x, b ~ c + d)

}
