% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\alias{stack.MultiFactor::MultiFactor}
\title{Combine levels across several LinkMaps in a MultiFactor}
\usage{
\method{stack}{`MultiFactor::MultiFactor`}(x, .path, out.format = c("LinkMap", "matrix"), ...)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.path}{a \code{formula} of length 2 with with levels to be merged separated
by the plus (\code{+}) sign. Optionally, a list with two character vectors,
signifying the variables to be combined at the left and right hand side,
respectively.}

\item{out.format}{\verb{Character scalar}. One of \code{'LinkMap'}, \code{'matrix'}.}

\item{...}{Additional arguments (unused.)}
}
\value{
a \code{LinkMap} or \verb{sparse Matrix}.
}
\description{
Generates a new \code{LinkMap} object by cross-referencing the elements of a
given \code{MultiFactor}. Elements can be merged by including several names,
separated by the plus (\code{+}) sign. See examples.
}
\examples{
x <- randomMultiFactor()
# Merge variables with "+" operator, new names get concatenated with ".":
stack(x, b ~ c + d)

}
