% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade_goods}
\alias{trade_goods}
\title{Small data set of six types of goods.}
\format{
\code{trade_goods}: A small tidy data set used in the vignette, describing
six items from six categories (36 rows); The 4 columns contain their
\code{name}, associated \code{emoji}, unicode \code{runes} (sans \verb{U+} prefix) and
\code{group} as well as their associated emoji and unicode ID, sans \verb{U+}
prefix.
}
\source{
\code{trade_goods}: Curated using the \code{emo} package.
}
\usage{
data("trade_goods", package = "MultiFactor")
}
\description{
Small data set of six types of goods.
}
\keyword{datasets}
