% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave-methods}
\alias{weave-methods}
\title{Weave a new LinkMap from a MultiFactor}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.path}{Either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{out.format}{\verb{Character scalar}. One of \code{'LinkMap'}, \code{'matrix'}.}

\item{include, exclude, exact}{\verb{Character vectors} Should feature types be
included or excluded from the available paths? Exact allows for exact
path definition.}
}
\value{
a \code{LinkMap} or \verb{sparse Matrix}.
}
\description{
Generates a new \code{LinkMap} object by cross-referencing the elements of a
given \code{MultiFactor}. Elements can be merged by including several names,
separated by the plus (\code{+}) sign. See examples.
}
\examples{
# Generate a random MultiFactor
x <- randomMultiFactor()

# Weave new b2c LinkMap
weave(x, b ~ c)
weave(x, b ~ a, out.format = "matrix")

}
