% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave}
\alias{weave}
\title{Weave a new LinkMap from a MultiFactor}
\usage{
weave(x, .by = NULL)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.by}{either a \code{formula} or a `character vector`` of length 2 with the
names of the desired combination of feature types.}
}
\value{
a \code{LinkMap}
}
\description{
Generates a new \code{LinkMap} object by cross-referencing the elements of a
given \code{MultiFactor}
}
\examples{
# Generate some random linkage input
a2b <- data.frame(
   a = sample(letters[seq(3)], 10, replace = TRUE),
   b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
a2c <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    c = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
x <- MultiFactor(list(a2b, a2c))
weave(x, b ~ c)
}
