% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave_all_paths}
\alias{weave_all_paths}
\title{Get and select paths to weave through a MutiFactor.}
\usage{
weave_all_paths(x, .by, as.paths = FALSE, include = NULL, exclude = NULL)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.by}{Either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{as.paths}{\verb{Boolean scalar} Whether to return an \code{igraph} path class, or
a list of character vectors (Default).}

\item{include, exclude}{\verb{Character vectors} Should feature types be included
or excluded from the available paths?}
}
\description{
Get and select paths to weave through a MutiFactor.
}
