% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave_apply.R
\name{weave_apply-methods}
\alias{weave_apply-methods}
\title{Index a table by a MultiFactor and apply arbitrary code to it}
\arguments{
\item{.x}{a \code{MultiFactor} object.}

\item{.path}{either a \code{formula} or a `character vector`` of length 2 with the
names of the desired combination of feature types.}

\item{.data}{\verb{A table}. A \code{data.frame}, \code{matrix}, other object with rows and
columns.}

\item{.fun}{A function, passed to lapply.}

\item{...}{Additional arguments passed to lapply call.}

\item{.index}{\verb{Character scalar} Where in \code{.x} can the target feature names be
found. (Default: "row.names")}
}
\value{
a \verb{Named list} of desired output.
}
\description{
\code{weave_apply} can run arbitrary code specified by \code{FUN} across any number
of row-based subsets of table \code{.data}. Mimics lapply.
}
\examples{

# Prepare data
x <- trade_posts()

# Generate small example feature table 'df'.
n <- nlevels(x)[["clothing"]]
df <- replicate(10, rbinom(n, rbinom(n, 100, runif(n)), runif(n)))

#' # Ensure rownames correspond to the second (RHS) variable in the formula.
df <- as.data.frame(df, row.names = levels(x)$clothing)

# Apply arbitrary code to x based on group membership
weave_apply(
    x,
    .path = fruit ~ clothing,
    .data = df,
    .fun = function(x) colSums(x)
)

}
\seealso{
\code{\link[=weave]{weave()}} \code{\link[=lapply]{lapply()}} \code{\link[=LinkMap]{LinkMap()}} \code{\link[=MultiFactor]{MultiFactor()}}
}
