% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave_to_tbl.R
\name{weave_to_tbl}
\alias{weave_to_tbl}
\title{Expand a table based on possibly overlapping group membership.}
\usage{
weave_to_tbl(x, .path, .data, .index = "row.names")
}
\arguments{
\item{x}{a \code{MultiFactor}.}

\item{.path}{either a \code{formula} or a `character vector`` of length 2 with the
names of the desired combination of feature types.}

\item{.data}{\verb{A table}. A \code{data.frame}, \code{matrix}, other object with rows and
columns. Should have a \code{cbind} method.}

\item{.index}{\verb{Character scalar} Column to look for feature IDs to link.
Default: "row.names".}
}
\value{
An expanded table \code{x}, with an added first column containing
subgroups.
}
\description{
Expand a table into a tidy table suitable for tidyverse-stype operations.
}
\examples{
# Prepare data
x <- trade_posts()

# Generate small example feature table 'df'.
n <- nlevels(x)[["clothing"]]
df <- replicate(10, rbinom(n, rbinom(n, 100, runif(n)), runif(n)))

#' # Ensure rownames correspond to the second (RHS) variable in the formula.
df <- as.data.frame(df, row.names = levels(x)$clothing)

# Apply arbitrary code to x based on group membership
weave_to_tbl(x, .path = fruit ~ clothing, .data = df)

}
