% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkMap-methods.R
\name{as.data.frame.LinkMap}
\alias{as.data.frame.LinkMap}
\alias{as.data.frame.MultiFactor::LinkMap}
\title{Convert a LinkMap to a data.frame
Convert a LinkMap back to a regular data.frame. Metadata is included.}
\usage{
\method{as.data.frame}{`MultiFactor::LinkMap`}(x, row.names, optional, ...)
}
\arguments{
\item{x}{a \code{LinkMap} object.}

\item{row.names, optional, ...}{For compatibility, not currently used.}
}
\value{
a \code{data.frame}
}
\description{
Convert a LinkMap to a data.frame
Convert a LinkMap back to a regular data.frame. Metadata is included.
}
\examples{
x <- randomLinkMap()
as.data.frame(x)
}
