% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{as.LinkMap}
\alias{as.LinkMap}
\title{Convert common classes to LinkMap}
\usage{
as.LinkMap(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{additional arguments}
}
\value{
a \code{LinkMap}
}
\description{
Convert common classes to LinkMap
}
\examples{
# Available methods:
as.LinkMap

}
\seealso{
\link{as.LinkMap-methods}
}
