% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkMap-methods.R
\name{as.matrix.LinkMap}
\alias{as.matrix.LinkMap}
\alias{as.matrix.MultiFactor::LinkMap}
\title{Convert a LinkMap to a sparse matrix.
Convert a LinkMap to a sparse matrix object from the \code{Matrix} package.}
\usage{
\method{as.matrix}{`MultiFactor::LinkMap`}(
  x,
  terms = colnames(x),
  dims = nlevels(x)[terms],
  dimnames = levels(x)[terms],
  value_id = NULL,
  force_pattern = is.null(value_id),
  ...
)
}
\arguments{
\item{x}{a \code{LinkMap} object.}

\item{terms}{id of cols in their desired order. \code{c(rows, cols)}.}

\item{dims}{length-2 integer vector of matrix dimensions.
Default: \code{colnames(x)}}

\item{dimnames}{list of dimnames. (Default: \code{levels(x)}).}

\item{value_id}{Name or index of column in metadata to use as matrix values.}

\item{force_pattern}{\code{Boolean}. Whether to ignore value and return a sparse
pattern matrix. (Default: FALSE)}

\item{...}{additional arguments. Not used.}
}
\value{
a sparse biadjacency \code{Matrix} with
}
\description{
Convert a LinkMap to a sparse matrix.
Convert a LinkMap to a sparse matrix object from the \code{Matrix} package.
}
\examples{
x <- randomLinkMap()
as.matrix(x)

}
\seealso{
\code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}}
}
