% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{as.MultiFactor}
\alias{as.MultiFactor}
\title{Convert common classes to MultiFactor}
\usage{
as.MultiFactor(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{additional arguments}
}
\value{
a \code{MultiFactor}.
}
\description{
Convert common classes to MultiFactor
}
\examples{
# Available methods:
as.MultiFactor


}
\seealso{
\link{as.LinkMap}
}
