% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{LinkMap}
\alias{LinkMap}
\title{LinkMap S7 container class}
\usage{
LinkMap(x, metadata = NULL)
}
\arguments{
\item{x}{\code{data.frame} with two named columns that can be coerced to factors.
Optionally, additional columns will be stored as metadata.}

\item{metadata}{Optional \code{data.frame} with same number of rows as x. Contains
information about the feature link in that row.}
}
\value{
a \code{LinkMap} object.
}
\description{
\code{LinkMap} is an S7 class to organize and manage multiple sets of factors,
for instance when tracing or converting feature IDs across databases. Methods
for \code{LinkMap} aim to follow \code{factor} behaviour.
}
\section{Slots}{

\describe{
\item{\code{levels}}{\verb{Named list} of character vectors depicting levels.}

\item{\code{metadata}}{\code{data.frame}. Optional. Used to store additional data or
application-specific tags.}
}}

\examples{
# Generate random linkage input
x <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    A = sample(LETTERS[seq(3)], 10, replace = TRUE)
)

# Create LinkMap
LinkMap(x)

}
\seealso{
\code{\link[=MultiFactor]{MultiFactor()}}
}
