% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{LinkMap}
\alias{LinkMap}
\title{LinkMap S7 container class}
\usage{
LinkMap(x)
}
\arguments{
\item{x}{a \code{data.frame} with two named columns that can be coerced to factors
and an optional third column of values associated with those two columns.}
}
\value{
a \code{LinkMap} object.
}
\description{
\code{LinkMap} is an S7 class to organize and manage multiple sets of factors,
for instance when tracing or converting feature IDs across databases. Methods
for \code{LinkMap} aim to follow \code{factor} behaviour.
}
\section{Slots}{

\describe{
\item{\code{levels}}{\verb{Named list} of character vectors depicting levels.}

\item{\code{value}}{\verb{Numeric vector} or \verb{boolean scalar}.}

\item{\code{is_bool}}{\code{Boolean}, returns \code{TRUE} if this object contains no \code{value}.}
}}

\examples{
# Generate random linkage input
x <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    A = sample(LETTERS[seq(3)], 10, replace = TRUE)
)

# Create LinkMap
LinkMap(x)

}
\seealso{
\code{\link[=MultiFactor]{MultiFactor()}}
}
