% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MultiFactor}
\alias{MultiFactor}
\title{MultiFactor S7 container class}
\usage{
MultiFactor(x)
}
\arguments{
\item{x}{a \code{LinkMap}, or named list of \code{LinkMap} objects.}
}
\value{
a \code{MultiFactor} object.
}
\description{
\code{MultiFactor} is an S7 class to organize and manage multiple sets of factors,
for instance when tracing or converting feature IDs across databases. Methods
for \code{MultiFactor} aim to follow \code{factor} behaviour.
}
\details{
The most straightforward way to construct a \code{MultiFactor} object is as a
named list of named data.frames. The columns of the data.frames indicate the
category of factor in that column.

A \code{MultiFactor} object presents itself similar to a \code{data.frame}, in the
sense that level types can be called as columns and individual data.frame
components can be called as rows.
\code{MultiFactor} inherits from \code{list}; Content can be accessed through regular
list methods (e.g., \code{[}, \code{[[}).
}
\section{Slots}{

\describe{
\item{\code{levels}}{\verb{Named list of character vectors}. Accessed through \code{levels(x)}}

\item{\code{value}}{Optional. \verb{Named list of vectors} of same size as \code{MultiFactor}
content, or \code{TRUE}, if missing.}

\item{\code{map}}{\verb{(sparse) Matrix} specifying which elements contain which levels.}
}}

\examples{
# Generate some random linkage input
a2b <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
a2c <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    c = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
x <- MultiFactor(list(a2b, a2c))

}
\seealso{
\code{\link[=MultiFactor-methods]{MultiFactor-methods()}}

\code{\link[=LinkMap]{LinkMap()}}
}
