% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor-methods.R
\name{MultiFactor-methods}
\alias{MultiFactor-methods}
\title{Methods for MultiFactor S7 container class}
\arguments{
\item{x, object}{\code{MultiFactor} on which the method should be applied.}
}
\value{
A MultiFactor
}
\description{
Methods for MultiFactor S7 container class
}
\examples{
# Setup
a2b <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
a2c <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    c = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
x <- MultiFactor(list(a2b, a2c))

# Basic properties
dim(x)
dimnames(x)

# Factor-like properties
levels(x)

# Retain MultiFactor structure using `[`.
x[1]

# Or extract individual LinkMaps using `[[`
x[["A"]]

# Combine using `c`:
c(x[2], x[1])

}
