% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor-methods.R
\name{MultiFactor-methods}
\alias{MultiFactor-methods}
\title{Methods for MultiFactor S7 container class}
\arguments{
\item{x, object}{\code{MultiFactor} on which the method should be applied.}
}
\value{
A MultiFactor
}
\description{
Methods for MultiFactor S7 container class
}
\examples{
# Setup
x <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    A = sample(LETTERS[seq(3)], 10, replace = TRUE)
) |> MultiFactor()
x

# Basic properties
dim(x)
dimnames(x)

# Factor-like properties
levels(x)

# Extract component LinkMaps using `[`.
x[1]
x["ec2cpd"]

}
