% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomMultiFactor.R
\name{randomMultiFactor}
\alias{randomMultiFactor}
\alias{randomLinkMap}
\title{Generate a random MultiFactor or LinkMap}
\usage{
randomMultiFactor(n_types = 6, n_features = 10, sparseness = 0.75)

randomLinkMap(x = list(lower = letters, UPPER = LETTERS), sparseness = 0.5)
}
\arguments{
\item{n_types}{\verb{Numeric scalar}, number of types of features to generate}

\item{n_features}{\verb{Numeric scalar}, number of features per type}

\item{sparseness}{\verb{Numeric scalar}, proportion: How rare are connections.
Default is \code{0.5}.}

\item{x}{optional list of two named vectors of features to use. Default is
\code{list(lower = letters, UPPER = LETTERS)}.}
}
\value{
a randomly generated object of the specified class.
}
\description{
Randomly generate a valid \code{MultiFactor} or \code{LinkMap} object.

called by \code{randomMultiFactor}, shouldn't be called by user.
}
\examples{
# Make a random MultiFactor object
randomMultiFactor()

# Make a random LinkMap object
randomLinkMap()

}
\seealso{
\code{\link[=MultiFactor]{MultiFactor()}}

\code{\link[=LinkMap]{LinkMap()}}
}
