% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playing_cards.R, R/randomMultiFactor.R
\name{draw_cards}
\alias{draw_cards}
\alias{poker_scores}
\alias{randomMultiFactor}
\alias{trade_posts}
\alias{randomLinkMap}
\title{Generate and optionally draw from a poker deck.}
\usage{
draw_cards(draw = NULL)

poker_scores()

randomMultiFactor(layout = NULL, n_features = 10, sparseness = 0.75)

trade_posts(raw.data = FALSE)

randomLinkMap(x = list(lower = letters, UPPER = LETTERS), sparseness = 0.5)
}
\arguments{
\item{draw}{Optional \code{integer}. How many cards to draw from the deck.}

\item{layout}{\code{igraph}, optional graph structure to generate random data for}

\item{n_features}{\verb{Numeric scalar}, number of features per type}

\item{sparseness}{\verb{Numeric scalar}, proportion: How rare are connections.
Default is \code{0.5}.}

\item{raw.data}{\code{Boolean}, Whether to return the \code{data.frame} of goods rather
than the default \code{MultiFactor}.}

\item{x}{optional list of two named vectors of features to use. Default is
\code{list(lower = letters, UPPER = LETTERS)}.}
}
\value{
a randomly generated object of the specified class.
}
\description{
Randomly generate a valid \code{MultiFactor} or \code{LinkMap} object.
\code{randomMultiFactor} can optionally take am \code{igraph} object to determine its
layout. (See examples)
\code{trade_posts()} generates a random \code{MultiFactor} in the style of the trading
example from the vignette.

called by \code{randomMultiFactor}, shouldn't be called by user.
}
\examples{
# Make a random MultiFactor object
randomMultiFactor()

# Use a (possibly random) igraph as input:
randomMultiFactor( igraph::sample_gnp(6, 2/3) )

# Make a random LinkMap object
randomLinkMap()

# Make a random MultiFactor with the trading goods from the vignettes
trade_posts()

# Playing cards
draw_cards(5)
poker_scores()

}
\seealso{
\link{playing_cards}

\code{\link[=MultiFactor]{MultiFactor()}}

\code{\link[=LinkMap]{LinkMap()}}
}
