% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_path.R
\name{select_path}
\alias{select_path}
\title{Define a path through a MultiFactor object.}
\usage{
select_path(
  x,
  .path,
  include = NULL,
  exclude = NULL,
  exact = NULL,
  as.edges = FALSE
)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.path}{Either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{include, exclude, exact}{\verb{Character vectors} Should feature types be
included or excluded from the available paths? Exact allows for exact
path definition.}

\item{as.edges}{\verb{Boolean scalar} Whether to return names of edges or nodes
(Default) in the path.}
}
\value{
a list of character vectors.
}
\description{
Define a path through a MultiFactor object.
}
\examples{
#' # Generate pair of random linkage input
a2b <- data.frame(
   a = sample(letters[seq(3)], 10, replace = TRUE),
   b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
a2c <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    c = sample(c("x", "y", "z"), 10, replace = TRUE)
)

# Create MultiFactor
x <- MultiFactor(list(a2b, a2c))

# Inspect a path between data types
select_path(x, b ~ c)
}
