% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{subgroup_to_tbl}
\alias{subgroup_to_tbl}
\title{Expand a table based on possibly overlapping group membership.}
\usage{
subgroup_to_tbl(x, link, .path, .index = "row.names")
}
\arguments{
\item{x}{\verb{A table}. A \code{data.frame}, \code{matrix}, other object with rows and
columns. Should have a \code{cbind} method.}

\item{link}{a \code{MultiFactor}.}

\item{.path}{either a \code{formula} or a `character vector`` of length 2 with the
names of the desired combination of feature types.}

\item{.index}{\verb{Character scalar} Column to look for feature IDs to link.
Default: "row.names".}
}
\value{
An expanded table \code{x}, with an added first column containing
subgroups.
}
\description{
Expand a table into a tidy table suitable for tidyverse-stype operations.
}
\examples{
# Prepare data
link <- trade_posts()

# Generate small example feature table 'x'.
n <- nlevels(link)[["clothing"]]
x <- replicate(10, rbinom(n, rbinom(n, 100, runif(n)), runif(n)))

#' # Ensure rownames correspond to the second (RHS) variable in the formula.
x <- as.data.frame(x, row.names = levels(link)$clothing)

# Apply arbitrary code to x based on group membership
subgroup_to_tbl(x, link, .path = fruit ~ clothing)

}
