% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{weave}
\alias{weave}
\title{Convert common classes to MultiFactor}
\usage{
weave(x, .path, ...)
}
\arguments{
\item{x}{input object}

\item{.path}{either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{...}{additional arguments}
}
\description{
Convert common classes to MultiFactor
}
\examples{
# Available methods:
weave

}
