% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{weave-generic}
\alias{weave-generic}
\alias{weave}
\title{Weave a path through an object}
\usage{
weave(x, .path, ...)
}
\arguments{
\item{x}{input object}

\item{.path}{either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{...}{additional arguments}
}
\value{
a \code{LinkMap} or \code{matrix}.
}
\description{
\code{weave()} is an S7 generic that finds a path through a
relational object with methods available for the following classes:
\itemize{
\item \code{MultiFactor::MultiFactor}
}
}
\examples{
# Available methods:
weave

}
