% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave.MultiFactor}
\alias{weave.MultiFactor}
\title{Weave a new LinkMap from a MultiFactor}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.by}{either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{out.format}{\verb{Character scalar}. One of \code{'LinkMap'}, \code{'matrix'}.}
}
\value{
a \code{LinkMap} or \verb{sparse Matrix}.
}
\description{
Generates a new \code{LinkMap} object by cross-referencing the elements of a
given \code{MultiFactor}
}
\examples{
# Generate pair of random linkage input
a2b <- data.frame(
   a = sample(letters[seq(3)], 10, replace = TRUE),
   b = sample(LETTERS[seq(3)], 10, replace = TRUE)
)
a2c <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    c = sample(c("x", "y", "z"), 10, replace = TRUE)
)

# Create MultiFactor
x <- MultiFactor(list(a2b, a2c))

# Weave new b2c LinkMap
weave(x, b ~ c)
weave(x, b ~ a, out.format = "matrix")
}
