% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{weave_apply-generic}
\alias{weave_apply-generic}
\alias{weave_apply}
\title{Index a table and apply arbitrary code to it}
\usage{
weave_apply(.x, .path, .data, .fun = NULL, ...)
}
\arguments{
\item{.x}{input relational object to dispatch on.}

\item{.path}{either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{.data}{an R object, such as tabular data.}

\item{.fun}{the function to be applied to each subgroup of \code{.x}.}

\item{...}{Optional arguments to \code{.fun}}
}
\value{
a list containing the results of \code{.fun}.
}
\description{
\code{weave_apply()} is an S7 generic that finds a path through a
relational object and evaluates provided code to each corresponding
subset of an input table.
}
\examples{
# Available methods:
weave_apply

}
