% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave-coverage.R
\name{weave_coverage}
\alias{weave_coverage}
\title{Perform enrichtment analysis from a weave}
\usage{
weave_coverage(
  x,
  .path,
  .data = NULL,
  metric = c("count", "coverage", "complete"),
  out.format = c("LinkMap", "matrix"),
  .data_column = "row.names"
)
}
\arguments{
\item{x}{a \code{MultiFactor}}

\item{.path}{Either a \code{formula} or a \verb{character vector} of length 2 with the
names of the desired combination of feature types.}

\item{.data}{Optional \verb{Character vector}. Lists observed features from the
found within the first element of \code{.path}. Alternatively, a \code{data.frame}
with the same information. (Also see \code{.data_column} argument).}

\item{metric}{\verb{Character scalar}.
One of \code{'count'}, \code{'coverage'}, \code{'complete'}.}

\item{out.format}{\verb{Character scalar}.
One of \code{'LinkMap'}, \code{'matrix'}.}

\item{.data_column}{\verb{Character scalar}. if \code{.data} is a table, where to find
feature IDs}
}
\value{
a \code{LinkMap} or \code{matrix} with the desired coverage information in
the @metadata slot.
}
\description{
Perform enrichtment analysis from a weave
}
\examples{
set.seed(2612)
# Draw five cards from a deck
drawn  <- draw_cards(5); drawn
scores <- poker_scores()

# Now enrich input
result <- weave_coverage(
    x = scores, .path = card ~ suit, .data = drawn, metric = "count"
)
# show result plus metadata
as.data.frame(result)

}
